perform simple unit conversions
Perform conversion based on unit tag. It ignores the
unit field on MetricType and uses the tag so that the
unit can be specified as part of the `tags` block on
a task. The unit for the MetricType may not be set and
may not have a fixed set of values.