add exclude config option
Adds an optional exclude setting that can be set
on the processor. This is a regex that will be
matched against the namespace string to drop
undesirable entries. This is mostly to handle
some cases where the wildcard '*' matches a bit
too much.

For example, psutil network interface stats
include an 'all' group. In Atlas we generally
don't want that because unless we restrict based
on the interface we'll get the aggregate over
all interfaces.